feat: user email과 비밀번호를 검증하는 미들웨어 추가
why: controller에서 직접 해결하는 방식보다 middleware를 거치는 방식으로
바꿨다. controller에서 작성해야 하는 logic의 수가 줄어들고, 또한 검증이
필요한 다른 controller에서도 middleware를 적용할 수 있어서 이 방식을
사용했다.
