+++ server/src/routes/channel/channel-router.ts
@@ -1,9 +1,9 @@
-import {Router} from "express";
+import { Router } from "express";
 import * as ChannelApiController from "../../controller/api/channel-controller";
 import PostRouter from "./post/post-router";
-import {isNumeric} from "../../validator/identifier-validator";
+import { isNumeric } from "../../middleware/validator";
 
-const router = Router({mergeParams: true});
+const router = Router({ mergeParams: true });
 
 /**
  *
@@ -11,13 +11,12 @@ const router = Router({mergeParams: true});
  * channel controller 의 find() 메소드 호출
  *
  * */
-router.route("/:title")
-        .get(ChannelApiController.find);
+router.route("/:title").get(ChannelApiController.find);
 
-
-router.route("/")
-        .get(ChannelApiController.findAll)
-        .post(ChannelApiController.create);
+router
+  .route("/")
+  .get(ChannelApiController.findAll)
+  .post(ChannelApiController.create);
 
 /**
  *
