+++ server/src/routes/user/user-router.ts
@@ -1,8 +1,9 @@
-import { Router, Request } from "express";
+import { Router } from "express";
 import * as UserApiController from "../../controller/api/user-controller";
+import { isValidUserForm } from "../../middleware/validator";
 
 const router = Router();
 
-router.post("/", UserApiController.create);
+router.post("/", isValidUserForm, UserApiController.create);
 router.get("/email/:email", UserApiController.findByEmail);
 export default router;
